Implement minimal photometry pipeline
