+++ src/mophongo/__init__.py
@@ -1,0 +1,15 @@
+"""Minimal photometry utilities."""
+
+from .psf import moffat_psf
+from .simulate import add_sources, simulate_image
+from .fit import fit_fluxes
+from .pipeline import run_photometry, PhotometryResult
+
+__all__ = [
+    "moffat_psf",
+    "add_sources",
+    "simulate_image",
+    "fit_fluxes",
+    "run_photometry",
+    "PhotometryResult",
+]
